+++ bindings/python/viet_bazi/_engine/index.js
@@ -1,4 +1,5 @@
 export { calculateAnnualTimeline, calculateBazi, ENGINE_VERSION, getMethodologyManifest } from './engine.js';
+export { BAZI_ERROR_CODES, BaziError, toBaziErrorPayload } from './errors.js';
 export { equationOfTime, solarCorrectionMinutes, solarLongitude, solarTermBoundary } from './calendar.js';
 export { BRANCHES, ELEMENTS, STEMS } from './constants.js';
 export { SHEN_SHA_CATALOG, SHEN_SHA_CODES } from './shen-sha.js';